added extention function to round
