+++ app/src/main/java/org/myspecialway/notifications/NotificationAlarmManager.kt
@@ -5,10 +5,9 @@ import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.provider.CalendarContract
 import org.myspecialway.common.Navigation
+import org.myspecialway.common.roundSeconds
 import org.myspecialway.ui.agenda.ScheduleRenderModel
-import java.util.*
 
 class NotificationAlarmManager(private val context: Context) {
     private var alarmManager: AlarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
@@ -26,16 +25,9 @@ class NotificationAlarmManager(private val context: Context) {
         alarmsQueue.add(pendingIntent)
         val triggerAtMillis = scheduleModel.time!!.date.time - System.currentTimeMillis()
 
-        val date = Date()
-        date.time = System.currentTimeMillis() + triggerAtMillis
-        val cal = Calendar.getInstance()
-        cal.time = date
-        cal.set(Calendar.SECOND, 0)
-        cal.set(Calendar.MILLISECOND, 0)
-
         alarmManager.set(
                 AlarmManager.RTC_WAKEUP,
-                cal.timeInMillis,
+                triggerAtMillis.roundSeconds(),
                 pendingIntent
         )
     }
